+++ trade-pa/android/app/src/main/java/uk/co/tradespa/app/MainActivity.java
@@ -1,0 +1,5 @@
+package uk.co.tradespa.app;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}